+++ runtime/lua/vim/_meta.lua
@@ -11,6 +11,7 @@ vim.diagnostic = require('vim.diagnostic')
 vim.filetype = require('vim.filetype')
 vim.fs = require('vim.fs')
 vim.func = require('vim.func')
+vim.glob = require('vim.glob')
 vim.health = require('vim.health')
 vim.highlight = require('vim.highlight')
 vim.iter = require('vim.iter')
